+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.1
+
+- Add custom expiry for single entry on write method
+- Add contains method
+
 ## 1.0.0
 
 - Initial version.